backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.1
pydantic[email]==2.5.1
asyncpg==0.29.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
redis==5.0.1

